# openapi.yaml
openapi: 3.1.0

# The spec for the endpoints in src/http.rs. It's written by hand rather
# than generated from the code, to avoid pulling in utoipa for five
# handlers. The catch is that it can fall out of date, so tests/fuzz_test.py
# checks a running binary against it.
#
# A movie doc stores every field OMDB returns, with snake_case keys. `Movie`
# only lists the fields this API guarantees; `additionalProperties: true`
# lets the rest (plot, actors, poster, ...) through, so a new OMDB field
# doesn't need a spec change.

info:
  title: moviedb
  version: 3.0.0
  description: |
    Self-hosted movie rating API. A movie's canonical address is
    `/movies/{imdb_id}`; title/year lookup is a filter on the collection.
    Every response, success or error, is JSON. Errors are RFC 9457
    problem-details objects served as `application/problem+json`.

    A method not listed here for a path gets a 405 with an `Allow` header.
  license:
    name: GPL-3.0

servers:
  - url: http://127.0.0.1:8123
    description: local test server (tests/smoke_test.py; fuzz_test.py passes its own --url)

security:
  - ApiKeyAuth: []

tags:
  - name: movies

paths:
  /movies:
    get:
      tags: [movies]
      summary: The movie collection, optionally filtered
      description: |
        Both filters are optional and are combined with AND. Leave both out
        to get every movie. An empty value such as `?year=` is treated as not
        given.

        Matching is exact and case-sensitive. Title and year together aren't
        unique, so a filter can match several movies or none. Either way the
        response is an array, possibly empty.
      operationId: listMovies
      parameters:
        - name: title
          in: query
          required: false
          description: Exact, case-sensitive title filter.
          schema:
            type: string
            examples: ["The Matrix"]
        - name: year
          in: query
          required: false
          description: Exact year filter. A string, matching how OMDB reports it.
          schema:
            type: string
            examples: ["1999"]
      responses:
        "200":
          description: The matching movies, as a bare array.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Movie"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
        "500": { $ref: "#/components/responses/InternalServerError" }

    post:
      tags: [movies]
      summary: Fetch a title from OMDB, store it, and snapshot its ratings
      description: |
        Looks the movie up on OMDB by title and year and stores it under its
        `imdb_id` with your rating, replacing any earlier copy. Every call
        also adds a `ratings_history` row for your rating and each of OMDB's,
        so to re-rate a movie, POST it again.

        The parameters go in the query string, not a request body. All three
        are required and can't be empty.
      operationId: addMovie
      parameters:
        - name: title
          in: query
          required: true
          schema:
            type: string
            minLength: 1
            examples: ["The Matrix"]
        - name: rating
          in: query
          required: true
          description: Your own score, returned as `personal`.
          schema:
            type: integer
            minimum: 0
            maximum: 100
            examples: [90]
        - name: year
          in: query
          required: true
          schema:
            type: string
            minLength: 1
            examples: ["1999"]
      responses:
        "200":
          description: |
            The movie was already stored. It has been updated and is
            returned. There's no `Location` header.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Movie"
        "201":
          description: A new movie was stored.
          headers:
            Location:
              description: The new movie's canonical URI, `/movies/{imdb_id}`.
              required: true
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Movie"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404":
          description: OMDB has no movie matching this title and year.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
        "500": { $ref: "#/components/responses/InternalServerError" }
        "502":
          description: |
            OMDB couldn't be reached, sent something that isn't JSON, returned
            an error this server doesn't recognise, or sent a movie with no
            `imdbID` or `Title`.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "503":
          description: OMDB's daily request quota is used up.
          headers:
            Retry-After:
              description: Always 86400, since OMDB doesn't say when the quota resets.
              required: true
              schema:
                type: string
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "504":
          description: OMDB didn't respond within 10 seconds.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"

  /movies/recent:
    get:
      tags: [movies]
      summary: Most-recently-refreshed movies, newest first
      description: |
        Sorted by `refreshed`, newest first. Like `/movies`, the response is
        a plain array.
      operationId: listRecent
      parameters:
        - name: limit
          in: query
          required: false
          description: |
            Defaults to 10. Anything above 50 is treated as 50. `limit=0`
            returns `[]`. The maximum below is just the largest number the
            server can parse; a bigger one is a 422.
          schema:
            type: integer
            minimum: 0
            maximum: 18446744073709551615
            default: 10
            examples: [8]
      responses:
        "200":
          description: The most recently refreshed movies, as a bare array.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Movie"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
        "500": { $ref: "#/components/responses/InternalServerError" }

  /movies/{imdb_id}:
    parameters:
      - $ref: "#/components/parameters/ImdbId"
    get:
      tags: [movies]
      summary: One movie
      description: |
        A movie's address, and where a 201 `Location` header points. This
        is the only endpoint that returns 404 for a missing movie.
      operationId: getMovie
      responses:
        "200":
          description: The stored movie doc.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Movie"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalServerError" }

  /movies/{imdb_id}/history:
    parameters:
      - $ref: "#/components/parameters/ImdbId"
    get:
      tags: [movies]
      summary: Ratings snapshots for one movie, oldest first
      description: |
        One entry per source per snapshot. Your rating appears with source
        `Personal`, recorded each time you POST; a refresh records OMDB's
        ratings. Before 3.0, every refresh also repeated your rating. Returns
        404 only if the movie is unknown. A known movie with no snapshots yet
        has an empty history.

        The migration from the old API created one snapshot per movie, dated
        at the time of the migration, but the ratings in it were fetched at
        some earlier, unrecorded date. Read the first snapshot for each movie
        as "correct as of the migration at the latest".
      operationId: getHistory
      responses:
        "200":
          description: The movie's ratings history.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/History"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalServerError" }

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: |
        Required on every request. The key is checked before routing, so an
        unknown path without a valid key gets a 401, not a 404.

  parameters:
    ImdbId:
      name: imdb_id
      in: path
      required: true
      description: The IMDb ID, e.g. `tt0133093`.
      schema:
        type: string
        pattern: "^tt[0-9]+$"
        # tests/fuzz_test.py adds this movie before it runs, so the examples
        # hit a real movie and not only 404s.
        examples: ["tt0133093"]

  schemas:
    Rating:
      type: object
      description: |
        One rating from OMDB. Values are strings in each source's own scale
        (`8.7/10`, `87%`, `73/100`) and aren't converted.
      required: [source, value]
      properties:
        source:
          type: string
          examples: ["Internet Movie Database", "Rotten Tomatoes", "Metacritic"]
        value:
          type: string
          examples: ["8.7/10"]
      additionalProperties: false

    Movie:
      type: object
      description: |
        OMDB's response with snake_case keys and its `Response` field
        dropped, plus two fields of ours: `personal` and `refreshed`. Only
        the fields this API guarantees are listed here; everything else from
        OMDB is passed through as is.
      required: [imdb_id, title, ratings, personal, refreshed]
      properties:
        imdb_id:
          type: string
          examples: ["tt0133093"]
        title:
          type: string
          examples: ["The Matrix"]
        year:
          type: string
          examples: ["1999"]
        imdb_rating:
          type: string
          examples: ["8.7"]
        box_office:
          type: string
          examples: ["$172,076,928"]
        ratings:
          type: array
          description: OMDB's ratings. Always present, possibly empty.
          items:
            $ref: "#/components/schemas/Rating"
        personal:
          type: integer
          minimum: 0
          maximum: 100
          description: Your score.
          examples: [90]
        refreshed:
          type: string
          description: |
            When the movie was last fetched from OMDB, by POST or by
            `refresh` (UTC, millisecond precision, `+00:00` suffix).
          examples: ["2026-07-17T12:34:56.789+00:00"]
      additionalProperties: true

    Snapshot:
      type: object
      required: [observed, source, value]
      properties:
        observed:
          type: string
          format: date-time
          description: ISO-8601 UTC timestamp, millisecond precision.
          examples: ["2026-08-14T09:12:44.317+00:00"]
        source:
          type: string
          examples: ["Internet Movie Database"]
        value:
          type: string
          examples: ["8.7/10"]
      additionalProperties: false

    History:
      type: object
      description: |
        An object rather than a plain array because it also includes the
        `imdb_id`.
      required: [imdb_id, snapshots]
      properties:
        imdb_id:
          type: string
          examples: ["tt0133093"]
        snapshots:
          type: array
          description: Oldest first, ties broken by source.
          items:
            $ref: "#/components/schemas/Snapshot"
      additionalProperties: false

    Problem:
      type: object
      description: An RFC 9457 problem-details object.
      required: [type, title, status, detail]
      properties:
        type:
          type: string
          description: Always `about:blank`; the status code says what went wrong.
          const: about:blank
        title:
          type: string
          description: The HTTP status text.
          examples: ["Not Found"]
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
          description: What went wrong this time.
          examples: ["Movie not found"]
      additionalProperties: false

  responses:
    BadRequest:
      description: Malformed path parameter (bad percent-escapes).
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"

    Unauthorized:
      description: Missing or invalid `x-api-key`.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"

    NotFound:
      description: Unknown `imdb_id`, or a path the API doesn't serve.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"

    UnprocessableEntity:
      description: A query parameter is missing, empty or can't be parsed.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"

    InternalServerError:
      description: Internal error; see the server logs (`journalctl -u moviedb`).
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"